### 2017-07-21 #327 @whedon
[PRE REVIEW]: UglyMol: a WebGL macromolecular viewer focused on the electron density
**Submitting author:** @wojdyr (<a href="http://orcid.org/0000-0003-3980-4092">Marcin Wojdyr</a>)
**Repository:** <a href="https://github.com/uglymol/uglymol" target ="_blank">https://github.com/uglymol/uglymol</a>
**Version:** v0.5.7
