### 2017-07-23 #327 @arokem
:hand: I can edit this one.

@wojdyr : do you have suggestions for reviewers? I haven't done crystallography in about 15 years, and don't know anyone in that field any more.
