### 2017-07-24 #327 @wojdyr
Thanks for looking into it.
I'm afraid I don't have suggestions for reviewers (people who come to my mind would have a conflict of interest).
But I think that a biochemist or a bioinformatician familiar with 3d structures could also be a good reviewer.
