### 2017-07-25 #327 @arokem
Hey @luizirber : would you be willing to review this paper for JOSS?
