### 2017-07-25 #327 @wojdyr
@arokem sure, I'll extend it. But are you sure about the recommendations? ATM the [author guidelines](http://joss.theoj.org/about#author_guidelines) show a 60-word example summary as the only hint about the required paper length.
