### 2017-07-25 #327 @arokem
This is undergoing change at the moment (if you want to see the discussion, it's going on over here: https://github.com/openjournals/joss/issues/294). We're working to use a better example. For the time being, take a look at https://github.com/cMadan/prism/blob/master/paper/paper.md
