### 2017-07-26 ae7037a
update a paper for openjournals/joss-reviews#327
