+++ paper.bib
@@ -55,3 +55,10 @@
   year = {submitted}
 }
 
+@online{molstack,
+  author = {Minor Lab, University of Virginia},
+  title = {molstack},
+  year = 2017,
+  url = {http://molstack.bioreproducibility.org}
+}
+
