+++ paper.md
@@ -24,21 +24,30 @@ macromolecular models together with the electron density.
 It uses web technologies (JavaScript and WebGL) and is suitable for embedding
 in web applications. The project was started as a fork of xtal.js [@xtaljs].
 
+Three-dimensional structural models of macromolecules are used to gain
+insights into biological processes. Most of the macromolecular structures
+are determined using X-ray crystallography, which provides information
+about electron density in a crystal. The electron density map is used to
+build a model and can be later used to check the local quality of the model.
+
 UglyMol is aimed at crystallographers who inspect electron density
 at various stages of structure solution and model completion.
 It can be also used by researchers who before using a model from
-the Protein Data Bank want to check how well the model is supported by the experimental data.
+the Protein Data Bank want to check how well the model is supported
+by the experimental data.
 
 To make UglyMol easy to use by its audience,
 the user interface is closely resembling Coot [@coot],
-a desktop program that is popular among crystallographers.
+a desktop program popular among crystallographers.
 
 UglyMol is currently used in a few web applications.
 In SynchWeb [@synchweb] in Diamond Light Source it presents results from
 the refinement pipeline Dimple [@dimple].
 In EXI [@exi] in European Synchrotron Radiation Facility it shows
 results from a phasing pipeline.
-And in CCP4 web services [@ccp4web] -- results from multiple programs
+In CCP4 web services [@ccp4web] -- results from multiple programs
 and pipelines.
+And in molstack [@molstack], which is a platforms for sharing electron
+density maps and their interpretations, it shows the shared maps.
 
 # References
