### 2017-07-26 #327 @wojdyr
I added a paragraph about scientific context.
And one sentence about another use of the software that I've just learnt about.

Let me know if you have other suggestions. I may reply with a delay -- I'm on holidays this week.
