### 2017-08-01 #327 @luizirber
@arokem: I am traveling indeed, just saw this issue. I can review it, but I will try to bring some people to help with the scientific discussion (since I don't know anything about crystallography 😅 )
