### 2017-08-04 #327 @arokem
OK - turns out I was broken. Closing in favor of #350 for the review.
