### 2017-08-04 #327 @arfon
> OK - turns out I was broken. Closing in favor of #350 for the review.

Whedon should also give smarter errors :-\
