### 2017-08-04 #327 @arokem
Now raised as an issue here: https://github.com/openjournals/joss/issues/306
