### 2017-09-06 #350 @luizirber
Hi! Sorry for the delay. A few comments:

## Version
- [ ] Tag latest version (0.5.7) on GitHub

## Community guidelines
- [ ] Install instructions could be listed in the README instead of only in the wiki
- [ ] "How to modify UglyMol?" could become the CONTRIBUTING file in GitHub

## Paper

- [ ] The last paragraph is a bit hard to read. Maybe make a markdown list?

I really liked the FAQ in the wiki! Lots of good info there.
