### 2017-09-18 #350 @arokem
@wojdyr : have you had a chance to take a look at these comments?
