### 2017-10-02 7f6021c
change rollup.config.js for Rollup 0.48
