### 2017-10-02 0a27086
add one line to the help
