### 2017-10-02 699683e
added CONTRIBUTING.md
