+++ CONTRIBUTING.md
@@ -1,0 +1,8 @@
+
+The FAQ has a section
+[https://github.com/uglymol/uglymol/wiki#how-to-modify-uglymol
+How to modify UglyMol].
+
+You are welcome to contribute your changes upstream.
+Open an Issue or Pull Request to discuss it.
+
