### 2017-10-02 a761489
changed the last paragraph of JOSS paper
