+++ paper.bib
@@ -62,3 +62,13 @@
   url = {http://molstack.bioreproducibility.org}
 }
 
+@article{contaminer,
+  author={Hungler, Arnaud and Momin, Afaque and Diederichs, Kay and Arold, Stefan T},
+  title={ContaMiner and ContaBase: a webserver and database for early identification of unwantedly crystallized protein contaminants},
+  journal={Journal of applied crystallography},
+  volume={49},
+  number={6},
+  pages={2252--2258},
+  year={2016},
+  publisher={International Union of Crystallography}
+}
