### 2017-10-02 #350 @wojdyr
Thank you for the review and sorry it took me a while to get back to it.

### Version
 yes, indeed, I didn't realize that I'm not pushing tags to the remote repository. I just pushed all the previous tags.

### Community guidelines:
* I keep the README file intentionally very short, hoping that it will make people more likely to read it.
* the CONTRIBUTING file is probably a good idea, I just added initial version of it to the repository

### Paper

I reworded the last paragraph.

Thanks for taking time to review the paper and the software.
