### 2017-10-08 #350 @arokem
@luizirber : does now meet all the requirements?
