### 2017-10-09 #350 @luizirber
Yup! I checked all the boxes, I think it is good to go!

💯 Thanks for the changes, @wojdyr
