### 2017-10-12 #350 @wojdyr
Thank you @luizirber and @arokem !
