### 2017-10-15 #350 @wojdyr
Thank you all!

The paper (both pdf and html) has mis-formatted author list in bibliography.
For example, this entry in the bib file:

    author = {Paul Emsley and Bernhard Lohkamp and William G. Scott and Kevin Cowtan}

is shown as:

    Emsley, Paul, Bernhard Lohkamp, William G. Scott, and Kevin Cowtan.

The first author (in each entry) has the first name second; other authors have the first name first.
