### 2017-10-15 #350 @arfon
@wojdyr - yeah, I've seen this before and to be honest I'm not exactly sure what's going on. If you're interested in digging into this the command we're using to compile the PDF (with Pandoc) is [here](https://github.com/openjournals/whedon/blob/master/lib/whedon/processor.rb#L155-L164).
